Enable DAT test
